samples/a2a-weather/requirements.txt
flask==3.0.0
requests==2.31.0
tavily-python>=0.3.0
python-dotenv==1.0.0
tiktoken>=0.7.0
langchain-google-genai==1.0.10
